src/main/webapp/app/entities/address/update/address-form.service.ts
import { Injectable } from '@angular/core';
import { FormControl, FormGroup, Validators } from '@angular/forms';

import { IAddress, NewAddress } from '../address.model';

/**
 * A partial Type with required key is used as form input.
 */
type PartialWithRequiredKeyOf<T extends { id: unknown }> = Partial<Omit<T, 'id'>> & { id: T['id'] };

/**
 * Type for createFormGroup and resetForm argument.
 * It accepts IAddress for edit and NewAddressFormGroupInput for create.
 */
type AddressFormGroupInput = IAddress | PartialWithRequiredKeyOf<NewAddress>;

type AddressFormDefaults = Pick<NewAddress, 'id'>;

type AddressFormGroupContent = {
  id: FormControl<IAddress['id'] | NewAddress['id']>;
  address1: FormControl<IAddress['address1']>;
  address2: FormControl<IAddress['address2']>;
  city: FormControl<IAddress['city']>;
  postcode: FormControl<IAddress['postcode']>;
  country: FormControl<IAddress['country']>;
  customer: FormControl<IAddress['customer']>;
};

export type AddressFormGroup = FormGroup<AddressFormGroupContent>;

@Injectable({ providedIn: 'root' })
export class AddressFormService {
  createAddressFormGroup(address: AddressFormGroupInput = { id: null }): AddressFormGroup {
    const addressRawValue = {
      ...this.getFormDefaults(),
      ...address,
    };
    return new FormGroup<AddressFormGroupContent>({
      id: new FormControl(
        { value: addressRawValue.id, disabled: true },
        {
          nonNullable: true,
          validators: [Validators.required],
        },
      ),
      address1: new FormControl(addressRawValue.address1, {
        validators: [Validators.required, Validators.maxLength(100)],
      }),
      address2: new FormControl(addressRawValue.address2, {
        validators: [Validators.maxLength(100)],
      }),
      city: new FormControl(addressRawValue.city, {
        validators: [Validators.required, Validators.maxLength(50)],
      }),
      postcode: new FormControl(addressRawValue.postcode, {
        validators: [Validators.required, Validators.maxLength(10), Validators.pattern('^[A-Za-z0-9\\s]+$')],
      }),
      country: new FormControl(addressRawValue.country, {
        validators: [Validators.required, Validators.maxLength(2)],
      }),
      customer: new FormControl(addressRawValue.customer),
    });
  }

  getAddress(form: AddressFormGroup): IAddress | NewAddress {
    return form.getRawValue() as IAddress | NewAddress;
  }

  resetForm(form: AddressFormGroup, address: AddressFormGroupInput): void {
    const addressRawValue = { ...this.getFormDefaults(), ...address };
    form.reset(
      {
        ...addressRawValue,
        id: { value: addressRawValue.id, disabled: true },
      } as any /* cast to workaround https://github.com/angular/angular/issues/46458 */,
    );
  }

  private getFormDefaults(): AddressFormDefaults {
    return {
      id: null,
    };
  }
}
